Move `build_next_js_app`
